Updates dataloader to "tapes" of independent dynamics runs
Previously, the data needed to be pre-concatenated into a single array.

Now, the data is stored in a directory of .npy files of independent dynamics runs.
The dataloader now concatenates them into a single array and stores the valid
starting indices of each tape. The getitem method now grabs the starting index
from the list of valid starting indices, but still returns the same tensors
(initial condition, targets, particle labels), and hopefully should slot in with
the rest of the codebase.